+++ examples/fully_connected_xor.js
@@ -4,7 +4,7 @@
  * Module dependencies
  */
 
-const {Net} = require('../models/net');
+const {Net} = require('../lib/net');
 const rectifiers = require('../lib/rectifiers');
 
 /**
