+++ lib/layer.js
@@ -4,7 +4,7 @@
  */
 
 const {Connection, ConnectionParams, Neuron} = require('./neuron');
-const errors = require('../lib/errors');
+const errors = require('./errors');
 
 /**
  * Module
