+++ lib/neuron.js
